refactored app to use Blueprints in routes architecture for better modularity
