+++ stock_api/blueprints/authentication/templates/login.html
@@ -29,7 +29,7 @@
 
         <div class="checkbox mb-3">
            <h6>Don't have an account?</h6>
-           <a class="btn btn-sm btn-secondary" href="{{ url_for('main.register') }}">Register</a>
+           <a class="btn btn-sm btn-secondary" href="{{ url_for('authentication.register') }}">Register</a>
         </div>
 
         {{ form.submit(class="btn btn-lg btn-block btn-primary") }}
