+++ stock_api/blueprints/authentication/templates/register.html
@@ -34,7 +34,7 @@
 
             <div class="checkbox mb-3">
               <h6>Already registered?</h6>
-              <a class="btn btn-sm btn-secondary" href="{{ url_for('main.login') }}">Login</a>
+              <a class="btn btn-sm btn-secondary" href="{{ url_for('authentication.login') }}">Login</a>
            </div>
 
             {{ form.submit(class="btn btn-lg btn-block btn-primary") }}
