+++ stock_api/blueprints/historical_data/templates/includes/line_chart_modal.html
