+++ stock_api/templates/base.html
@@ -26,7 +26,7 @@
     <!-- NavBar Start -->
     <div id="navbar">
       <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-        <a class="navbar-brand" href="{{ url_for('main.home') }}">Stokk.io</a>
+        <a class="navbar-brand" href="{{ url_for('historical_data.home') }}">Stokk.io</a>
         <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
         </button>
@@ -52,16 +52,16 @@
               <a class="nav-link disabled">{{ current_user.username }}</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="{{ url_for('main.logout') }}">Logout</a>
+              <a class="nav-link" href="{{ url_for('authentication.logout') }}">Logout</a>
             </li>
           </ul>
           {% else %}
           <ul class="navbar-nav">
             <li class="nav-item">
-              <a class="nav-link" href="{{ url_for('main.login') }}">Login</a>
+              <a class="nav-link" href="{{ url_for('authentication.login') }}">Login</a>
             </li>
             <li class="nav-item">
-              <a class="nav-link" href="{{ url_for('main.register') }}">Register</a>
+              <a class="nav-link" href="{{ url_for('authentication.register') }}">Register</a>
             </li>
           </ul>
           {% endif %}
